+++ CHANGELOG.md
@@ -6,6 +6,10 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ## unreleased
 
+### Changed
+
+* Update to jsonwebtoken 3.
+
 ### Fixed
 
 * Remove compiler warnings